combine multiple data types into single file
